Send key strokes to snap windows
This commit sends key strokes to snap windows instead of setting an explicit
layout using window sizes and positions. The latter only works for one
resolution whereas the former is resolution independent.

The key strokes used are Super_L+Right to snap the current window (with Emacs)
to the right and Super_L+Left to snap a given window to the left. This only
works if these keys are bound to these window layouts.