Only tile windows that don't seem snapped
In Xfce, if you send the keystroke to snap an X window to a window that is
already snapped, it toggles the snap. This commit only sends the key stroke if
the window doesn't appear to be snapped yet.

This commit does a bit more than what the subject line says, for example it
fixes the snap functions, it refactors the code to determine what to snap,
renames functions etc. Sorry about that...